Flytt byggBegrunnelserOgFritekster inn i BrevPeriodeGenerator
