+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/brev/domene/BrevPeriodeGrunnlagMedPersoner.kt
@@ -1,11 +1,5 @@
 package no.nav.familie.ba.sak.kjerne.brev.domene
 
-import no.nav.familie.ba.sak.common.NullablePeriode
-import no.nav.familie.ba.sak.kjerne.behandlingsresultat.MinimertUregistrertBarn
-import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.Målform
-import no.nav.familie.ba.sak.kjerne.vedtak.domene.Begrunnelse
-import no.nav.familie.ba.sak.kjerne.vedtak.domene.FritekstBegrunnelse
-import no.nav.familie.ba.sak.kjerne.vedtak.domene.tilBrevBegrunnelse
 import no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode.Vedtaksperiodetype
 import java.time.LocalDate
 
@@ -17,27 +11,4 @@ data class BrevPeriodeGrunnlagMedPersoner(
     val fritekster: List<String> = emptyList(),
     val minimerteUtbetalingsperiodeDetaljer: List<MinimertUtbetalingsperiodeDetalj> = emptyList(),
     val erFørsteVedtaksperiodePåFagsak: Boolean,
-) {
-    fun byggBegrunnelserOgFritekster(
-        restBehandlingsgrunnlagForBrev: RestBehandlingsgrunnlagForBrev,
-        uregistrerteBarn: List<MinimertUregistrertBarn> = emptyList(),
-        brevMålform: Målform,
-    ): List<Begrunnelse> {
-
-        val brevBegrunnelser = this.begrunnelser.sorted()
-            .map {
-                it.tilBrevBegrunnelse(
-                    vedtaksperiode = NullablePeriode(this.fom, this.tom),
-                    personerIPersongrunnlag = restBehandlingsgrunnlagForBrev.personerPåBehandling,
-                    brevMålform = brevMålform,
-                    uregistrerteBarn = uregistrerteBarn,
-                    minimerteUtbetalingsperiodeDetaljer = this.minimerteUtbetalingsperiodeDetaljer,
-                    minimerteRestEndredeAndeler = restBehandlingsgrunnlagForBrev.minimerteEndredeUtbetalingAndeler
-                )
-            }
-
-        val fritekster = fritekster.map { FritekstBegrunnelse(it) }
-
-        return brevBegrunnelser + fritekster
-    }
-}
+)
